General tidy up. No significant code changes.
