+++ lib/settings_page_contents.dart
@@ -2,10 +2,10 @@
 import 'package:flutter/material.dart';
 
 //  Import project-specific files.
-import 'package:kar_kam/app_model.dart';
-import 'package:kar_kam/app_settings_data.dart';
+import 'package:kar_kam/old_app_settings_data.dart';
 import 'package:kar_kam/lib/get_it_service.dart';
 import 'package:kar_kam/lib/data_store.dart';
+import 'package:kar_kam/settings.dart';
 import 'package:kar_kam/settings_page_list_tile.dart';
 
 /// [SettingsPageContents] provides the settings page PageSpec contents.
@@ -93,13 +93,13 @@ class _SettingsPageContentsState extends State<SettingsPageContents> {
         ),
         onTap: () {
           //  Toggle bool variable in AppModel that controls the fade effect.
-          GetItService.instance<AppModel>().toggleDrawLayoutBounds();
+          GetItService.instance<Settings>().toggleDrawLayoutBounds();
         },
         trailing: Icon(
           Icons.circle_notifications_outlined,
           size: AppSettingsOrig.settingsPageListTileIconSize,
         ),
-        widget: Text(
+        widget: const Text(
           '5. Click to switch drawLayoutBounds',
           maxLines: 1,
           softWrap: false,
@@ -117,13 +117,13 @@ class _SettingsPageContentsState extends State<SettingsPageContents> {
         ),
         onTap: (() {
           //  Toggle bool variable in AppModel that controls the fade effect.
-          GetItService.instance<AppModel>().toggleSettingsPageListTileFadeEffect();
+          GetItService.instance<Settings>().toggleSettingsPageListTileFadeEffect();
         }),
         trailing: Icon(
           Icons.circle_notifications_outlined,
           size: AppSettingsOrig.settingsPageListTileIconSize,
         ),
-        widget: Text(
+        widget: const Text(
           '6. Click to toggle settingsPageListTileFadeEffect!',
           maxLines: 1,
           softWrap: false,
